chunk abstraction
I've begun to split chunk logic into multiple layers. My current plan is to have the following additional controllers:

- ChunkTerrainController
-- Generates base terrain

- ChunkBlocksController
-- Handles block changes and updates

- ChunkMeshController
-- Handles chunk meshing

Hopefully these abstractions will make chunk logic a lot simpler.

I am also considering a way to maintain the threaded nature of chunk generation, but not initializing / retrieving a cached job until the job is being worked on. This would allow the cache to self-size, and not take up a load of memory as dozens of chunks queue up for generation and instantiate new jobs.