add MainThreadActionController
MainThreadActionController allows actions to be added to a stack and executed every frame as frame time allows. This is for executing things on the main thread that would otherwise not be able to be executed there because of control flow.

The controller is agnostic of external control flow, only executing given commands, so any parameters or state information must be handled within the action passed.

Also, for some reason, everything seems to work correctly. At least the generation. Marvellous.