optimize OctreeNode<T>
Significant CPU time is spent in the `.SetPoint()` function. This had caused me to take a close look at each function call within, in an effort to slim any excess. Things are less transparent now as far as functionality, but it should also be more performant, which in such a critical context is preferable.