performant equivalence check
Apparently, by the grace of Microsoft, `where T : unmanaged` in tandem with `.Equals()` doesn't use any kind of numeric check, but instead uses an `object.Equals(object)` check. Or at least, it sure seems that way given that checking equivalence of unmanaged types with `.GetHashCode()` results in a 4-5x decrease in execution time. So that's nice.